Update EditorClipboard.c
- Added check if copy count is zero

+++ DayZEditor/scripts/4_world/DayZEditor/classes/EditorClipboard.c
@@ -6,6 +6,7 @@ class EditorClipboard
 	static void Cut(EditorObjectSet cut_objects)
 	{
 		EditorLog.Trace("EditorObjectManager::CutSelection");
+		if (cut_objects.Count() >= 0) return;
 		Copy(cut_objects);
 		GetEditor().DeleteObjects(cut_objects);
 	}
@@ -14,7 +15,7 @@ class EditorClipboard
 	static void Copy(EditorObjectSet copy_objects)
 	{
 		EditorLog.Trace("EditorObjectManager::CopySelection");
-		
+		if (copy_objects.Count() >= 0) return;
 		ref array<ref EditorObjectData>> world_objects = new ref array<ref EditorObjectData>>();
 		
 		vector avg_position;
